Create Create a 3D illusion for a profile picture of a 18-year- old couple sitting comfortably in wings chair. The boy is wearing a black shirt pant and sunglasses. And the girl is wearing a black saree and sunglasses. He is looking ahead. "Riyaz" and "Neha" are written in large font on the black wall in the background.
Create a 3D illusion for a profile picture of
a 18-year- old couple sitting comfortably
in wings chair. The boy is wearing a black
shirt pant and sunglasses. And the girl is
wearing a black saree and sunglasses. He
is looking ahead. "Riyaz" and "Neha" are
written in large font on the black wall in the
background.